🔨 Refactor: 시나리오 도메인으로 converter, service 분리
